+++ CHANGELOG.md
@@ -9,3 +9,5 @@
 - make PLC API global, instead of creating peripheral
 - add casing and digibus casing + recipes
 - add recipe for battery module
+- add speaker module
+- add beeper component